feat(trading): extract isEip7702DelegatedAccount, plumb signing overrides
Move the helper into utils/ and re-export from the trading package
index so it can be used outside postCoWProtocolTrade.

Pass env + settlementContractOverride to OrderSigningUtils.getDomain
on the 7702 branch so override settings actually take effect for
delegated accounts.

For 65-byte ECDSA output from a 7702 delegate, honor an explicit
EIP1271 signingScheme by wrapping via the (order, sig) ABI tuple
instead of forcing EIP712.
