+++ packages/trading/src/index.ts
@@ -22,4 +22,5 @@ export { getEthFlowContract } from './getEthFlowTransaction'
 export * from './appDataUtils'
 export * from './calculateUniqueOrderId'
 export { getPartnerFeeBps } from './utils/getPartnerFeeBps'
+export { isEip7702DelegatedAccount } from './utils/isEip7702DelegatedAccount'
 export { mapQuoteAmountsAndCosts, swapParamsToLimitOrderParams, getTradeParametersAfterQuote } from './utils/misc'
